Move Box class to utilities
